server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai-whisper==20231117
requests==2.31.0
python-multipart==0.0.6
pynacl==1.5.0
python-dotenv==1.0.0
keyring==24.3.0
